✨ feat: Implement layout for multiplayer game room list page
